Completed error handling during file transfer. Multiple files can be transferred now. media_clipboard now has bounce effect and dismiss functionality. Persistent login implemented.
The Andorid and Windows versions will deviate here for me to work seperately on the different UI and Advertisement support.

The app is stable enough now, so this version will be pulled into the main branch.

All subsequent platform-independent enhancements will go through the beta branch to the main branch.
While two new branches will be created 'Windows' and 'Android' that will have the contents of the main branch along with platform specific functionality and enhancements.